Fix deadlock between Executor and UserTaskManager (#2086)
This PR fixes the deadlock introduced by PR #2069 by partially reverting changes in that PR by moving _userTaskManager.markTaskExecutionBegan(_uuid) from the constructor of ProposalExecutionRunnable back to its run() thread. This avoids the deadlock formed between Executor and UserTaskManager. The initialization of ProposalExecutionRunnable is under Executor's monitor lock and _userTaskManager.markTaskExecutionBegan aquires UserTaskManager's monitor lock. On the other hand, the process of StopProposalRequest enters UserTaskManager's monitor lock first, and then acquires Executor's monitor lock.

With this change, we would have the issue back that [request]-successful-request-execution-timer may or may not record the proposal computing time for request that requires proposal execution depending on the race condition between the task is removed from _uuidToActiveUserTaskInfoMap and checkActiveUserTasks is called. This issue will be addressed in a separate PR.